Properly implement phase 3.
The source file is decomposed into preprocessing tokens6) and sequences
of white-space characters (including comments). A source file shall not
end in a partial preprocessing token or in a partial comment. Each comment is
replaced by one space character. New-line characters are retained. Whether each
nonempty sequence of white-space characters other than new-line is retained or
replaced by one space character is implementation-defined.

@r550:

--- FAIL: TestPreprocessor (38.40s)
	all_test.go:493: [testdata/arith-1.c]
	all_test.go:493: [testdata/dev/sqlite3/shell.c]
	all_test.go:493: [testdata/dev/sqlite3/sqlite3.c]
	all_test.go:493: [testdata/dev/sqlite3/sqlite3.h]
	all_test.go:493: [testdata/dev/sqlite3/sqlite3ext.h]
	all_test.go:493: [testdata/dev/vim/vim/src/buffer.c]
	all_test.go:493: [testdata/dev/vim/vim/src/blowfish.c]
	all_test.go:493: [testdata/dev/vim/vim/src/charset.c]
	all_test.go:493: [testdata/dev/vim/vim/src/crypt.c]
	all_test.go:493: [testdata/dev/vim/vim/src/crypt_zip.c]
	all_test.go:493: [testdata/dev/vim/vim/src/diff.c]
	all_test.go:493: [testdata/dev/vim/vim/src/digraph.c]
	all_test.go:493: [testdata/dev/vim/vim/src/edit.c]
	all_test.go:493: [testdata/dev/vim/vim/src/eval.c]
	all_test.go:493: [testdata/dev/vim/vim/src/ex_cmds.c]
	all_test.go:493: [testdata/dev/vim/vim/src/ex_cmds2.c]
	all_test.go:493: [testdata/dev/vim/vim/src/ex_docmd.c]
	all_test.go:493: [testdata/dev/vim/vim/src/ex_eval.c]
	all_test.go:493: [testdata/dev/vim/vim/src/ex_getln.c]
	all_test.go:568: [testdata/dev/vim/vim/src/ex_getln.c]: got
87672 tokens, expected 87696 tokens (∆ -24)
	all_test.go:581: 84319
		got testdata/dev/vim/vim/src/ex_getln.c:6022:9:
IDENTIFIER "STRCAT"
		exp /tmp/cc-test-383486696:20708:18: IDENTIFIER "strcat"
	all_test.go:493: [testdata/dev/vim/vim/src/fileio.c]
	all_test.go:493: [testdata/dev/vim/vim/src/fold.c]
	all_test.go:493: [testdata/dev/vim/vim/src/getchar.c]
	all_test.go:493: [testdata/dev/vim/vim/src/hardcopy.c]
	all_test.go:493: [testdata/dev/vim/vim/src/hashtab.c]
	all_test.go:493: [testdata/dev/vim/vim/src/if_cscope.c]
	all_test.go:493: [testdata/dev/vim/vim/src/if_xcmdsrv.c]
	all_test.go:493: [testdata/dev/vim/vim/src/mark.c]
	all_test.go:493: [testdata/dev/vim/vim/src/memline.c]
	all_test.go:493: [testdata/dev/vim/vim/src/menu.c]
	all_test.go:493: [testdata/dev/vim/vim/src/message.c]
	all_test.go:493: [testdata/dev/vim/vim/src/misc1.c]
	all_test.go:493: [testdata/dev/vim/vim/src/misc2.c]
	all_test.go:493: [testdata/dev/vim/vim/src/move.c]
	all_test.go:493: [testdata/dev/vim/vim/src/mbyte.c]
	all_test.go:493: [testdata/dev/vim/vim/src/normal.c]
	all_test.go:493: [testdata/dev/vim/vim/src/ops.c]
	all_test.go:493: [testdata/dev/vim/vim/src/option.c]
	all_test.go:493: [testdata/dev/vim/vim/src/os_unix.c]
	all_test.go:493: [testdata/dev/vim/vim/src/auto/pathdef.c]
	all_test.go:493: [testdata/dev/vim/vim/src/popupmnu.c]
	all_test.go:493: [testdata/dev/vim/vim/src/quickfix.c]
	all_test.go:493: [testdata/dev/vim/vim/src/regexp.c]
	all_test.go:493: [testdata/dev/vim/vim/src/screen.c]
	all_test.go:493: [testdata/dev/vim/vim/src/search.c]
	all_test.go:493: [testdata/dev/vim/vim/src/sha256.c]
	all_test.go:493: [testdata/dev/vim/vim/src/spell.c]
	all_test.go:493: [testdata/dev/vim/vim/src/syntax.c]
	all_test.go:493: [testdata/dev/vim/vim/src/tag.c]
	all_test.go:493: [testdata/dev/vim/vim/src/term.c]
	all_test.go:493: [testdata/dev/vim/vim/src/ui.c]
	all_test.go:493: [testdata/dev/vim/vim/src/undo.c]
	all_test.go:493: [testdata/dev/vim/vim/src/window.c]
	all_test.go:493: [testdata/dev/vim/vim/src/netbeans.c]
	all_test.go:493: [testdata/dev/vim/vim/src/channel.c]
	all_test.go:493: [testdata/dev/vim/vim/src/json.c]
	all_test.go:493: [testdata/dev/vim/vim/src/main.c]
	all_test.go:493: [testdata/dev/vim/vim/src/memfile.c]
	all_test.go:493: [testdata/dev/vim/vim/src/version.c]
	all_test.go:568: [testdata/dev/vim/vim/src/version.c]: got 64073
tokens, expected 64069 tokens (∆ 4)
	all_test.go:581: 58727
		got testdata/dev/vim/vim/src/version.c:52:21:
STRINGLITERAL "\"VIM - Vi IMproved 7.4 (2013 Aug 10, compiled \""
		exp /tmp/cc-test-708392661:15190:21: STRINGLITERAL
"\"VIM - Vi IMproved 7.4 (2013 Aug 10, compiled Mar 10 2016 18:48:03)\""
FAIL
exit status 1
FAIL	github.com/cznic/cc	38.518s

+++ all_test.go
@@ -70,6 +70,7 @@ var (
 	oFailFast = flag.Bool("ff", false, "crash on first reported error (in some tests.)")
 	oLong     = flag.Bool("long", false, "Enable long tests. (false)")
 	oRE       = flag.String("re", "", "Regexp filter.")
+	oTmp      = flag.Bool("tmp", false, "keep certain temp files.")
 
 	includes = []string{}
 
@@ -489,6 +490,7 @@ func testPreprocessor(t *testing.T, predefine string, cppOpts, src []string, opt
 		}
 	}
 
+	t.Log(src)
 	logf, err := os.Create("log-" + filepath.Base(src[len(src)-1]))
 	if err != nil {
 		t.Fatal(err)
@@ -502,7 +504,7 @@ func testPreprocessor(t *testing.T, predefine string, cppOpts, src []string, opt
 
 	var got, exp []xc.Token
 
-	if _, err := Parse(
+	_, err = Parse(
 		predefine,
 		src,
 		newTestModel(),
@@ -519,8 +521,9 @@ func testPreprocessor(t *testing.T, predefine string, cppOpts, src []string, opt
 			}),
 			disableWarnings(),
 		)...,
-	); err != nil {
-		t.Error(err)
+	)
+	if err != nil {
+		t.Error(errString(err))
 		return
 	}
 
@@ -536,7 +539,9 @@ func testPreprocessor(t *testing.T, predefine string, cppOpts, src []string, opt
 	}
 
 	defer func() {
-		os.Remove(f.Name())
+		if !*oTmp {
+			os.Remove(f.Name())
+		}
 		f.Close()
 	}()
 
@@ -579,7 +584,7 @@ func testPreprocessor(t *testing.T, predefine string, cppOpts, src []string, opt
 	}
 }
 
-func dirExists(dir string) bool {
+func dirExists(t *testing.T, dir string) bool {
 	dir = filepath.FromSlash(dir)
 	fi, err := os.Stat(dir)
 	if err != nil {
@@ -587,11 +592,11 @@ func dirExists(dir string) bool {
 			return false
 		}
 
-		panic(err)
+		t.Fatal(err)
 	}
 
 	if !fi.IsDir() {
-		panic(err)
+		t.Fatal(err)
 	}
 
 	return true
@@ -614,7 +619,7 @@ func TestPreprocessor(t *testing.T) {
 
 	testPreprocessor(t, predefined, nil, []string{"testdata/arith-1.c"})
 
-	if !dirExists("testdata/dev/sqlite3") {
+	if !dirExists(t, "testdata/dev/sqlite3") {
 		return
 	}
 
@@ -623,7 +628,7 @@ func TestPreprocessor(t *testing.T) {
 	testPreprocessor(t, predefined, nil, []string{"testdata/dev/sqlite3/sqlite3.h"}, sysIncludes)
 	testPreprocessor(t, predefined, nil, []string{"testdata/dev/sqlite3/sqlite3ext.h"}, sysIncludes)
 
-	if !dirExists("testdata/dev/vim") {
+	if !dirExists(t, "testdata/dev/vim") {
 		return
 	}
